refactor: API Service 디렉토리 변경
#43
